Refactor jovian.py into a modular package
- Split jovian.py into moons.py, mutual.py, grs.py, and utils.py.
- Extract is_inside_ellipsoid_projection into utils.py.
- Update internal imports to use relative paths.
- Update unit tests to mock get_timescale in the new mutual submodule.
- Maintain public API via __init__.py.
